requirements.txt
pycryptodome>=3.19.0
sympy>=1.12
pytest>=7.0.0